@@ -2,36 +2,111 @@
   "name": "vcl2",
   "private": true,
   "scripts": {
-    "start": "meteor run"
+    "start": "meteor run",
+    "lint": "eslint --ext .js,.jsx .",
+    "pretest": "npm run lint --silent"
   },
   "dependencies": {
     "@babel/runtime": "^7.0.0-beta.46",
     "aws-sdk": "^2.252.1",
-    "babel-runtime": "^6.26.0",
     "bcrypt": "^1.0.3",
     "chai": "^4.1.2",
     "core-js": "^2.5.6",
-    "eslint": "^4.19.1",
     "eslint-config-react-tools": "^1.2.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
     "guacamole-common-js": "^0.9.14-b",
-    "history": "^4.7.2",
+    "guacamole-lite": "^0.6.2",
     "meteor-node-stubs": "^0.2.11",
-    "normalize.css": "^8.0.0",
     "prop-types": "^15.6.1",
-    "react": "^16.3.2",
-    "react-addons-css-transition-group": "^15.6.2",
+    "react": "^16.4.1",
     "react-bootstrap": "^0.32.1",
     "react-bootstrap-sidebar": "0.0.1",
     "react-bootstrap-table": "^4.3.1",
-    "react-dom": "^16.3.2",
-    "react-icons-kit": "^1.1.4",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
+    "react-dom": "^16.4.1",
+    "react-icons-kit": "^1.1.5",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "^1.1.1",
     "react-sidebar": "^2.3.2",
     "react-sidenav": "^2.1.3",
     "react-table": "^6.8.0",
-    "simpl-schema": "^1.5.0"
+    "react-transition-group": "^2.3.1"
+  },
+  "devDependencies": {
+    "@meteorjs/eslint-config-meteor": "^1.0.5",
+    "babel-eslint": "^8.2.3",
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-import-resolver-meteor": "^0.4.0",
+    "eslint-plugin-import": "^2.12.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-meteor": "^5.0.0",
+    "eslint-plugin-react": "^7.9.1"
+  },
+  "eslintConfig": {
+    "parser": "babel-eslint",
+    "parserOptions": {
+      "ecmaVersion": 6,
+      "sourceType": "module",
+      "ecmaFeatures": {
+        "jsx": true
+      },
+      "allowImportExportEverywhere": true
+    },
+    "plugins": [
+      "meteor"
+    ],
+    "extends": [
+      "airbnb",
+      "plugin:meteor/recommended"
+    ],
+    "rules": {
+      "import/extensions": "off",
+      "import/no-extraneous-dependencies": "off",
+      "import/prefer-default-export": "off",
+      "no-underscore-dangle": "off",
+      "class-methods-use-this": "off",
+      "object-shorthand": [
+        "error",
+        "always",
+        {
+          "avoidQuotes": false
+        }
+      ],
+      "meteor/eventmap-params": [
+        "error",
+        {
+          "eventParamName": "event",
+          "templateInstanceParamName": "instance"
+        }
+      ],
+      "meteor/template-names": [
+        "off"
+      ],
+      "react/jsx-filename-extension": [
+        1,
+        {
+          "extensions": [
+            ".js",
+            ".jsx"
+          ]
+        }
+      ],
+      "react/forbid-prop-types": "off",
+      "jsx-a11y/no-static-element-interactions": "off",
+      "jsx-a11y/click-events-have-key-events": "off",
+      "jsx-a11y/no-noninteractive-element-interactions": "off",
+      "jsx-a11y/anchor-is-valid": "off"
+    },
+    "settings": {
+      "import/resolver": "meteor"
+    }
+  },
+  "postcss": {
+    "plugins": {
+      "autoprefixer": {
+        "browsers": [
+          "last 2 versions"
+        ]
+      }
+    }
   }
 }
