@@ -1,8 +1,12 @@
 import { Meteor } from 'meteor/meteor';
+import React from 'react';
 import { render } from 'react-dom';
-import { renderRoutes } from '../imports/startup/client/routes.jsx';
-import '../imports/startup/client/accounts-config.js';
+
+import AppContainer from '../imports/ui/containers/AppContainer.jsx';
 
 Meteor.startup(() => {
-  render(renderRoutes(), document.getElementById('app'));
+  render(
+    <AppContainer />,
+    document.getElementById('app'),
+  );
 });
