@@ -1,19 +1,14 @@
 import { Meteor } from 'meteor/meteor';
-import { Session } from 'meteor/session';
+import { ReactiveVar } from 'meteor/reactive-var';
 import { withTracker } from 'meteor/react-meteor-data';
 
-import App from '../layouts/App.jsx'
-import { Tasks } from '../../api/events.js';
+import App from '../layouts/App.jsx';
 
-export default withTracker(() => {
-  return {
-    user: Meteor.user(),
-    connected: Meteor.status().connected,
-    menuOpen: Session.get('menuOpen'),
-    tasks: Tasks.find({}).fetch(),
-    /*lists: Lists.find({ $or: [
-      { userId: { $exists: false } },
-      { userId: Meteor.userId() },
-    ] }).fetch(), */
-  };
-})(App);
+const menuOpen = new ReactiveVar(false);
+
+export default withTracker(() => ({
+  user: Meteor.user(),
+  loading: false,
+  connected: Meteor.status().connected,
+  menuOpen,
+}))(App);
